refactor: Move cache sync logic to generator/merger/sync.py
